Sync.
git-svn-id: https://svn.xiph.org/trunk/ezstream@13460 0101bb08-14d6-0310-b084-bc0e0c8e3800

+++ src/xalloc.c
@@ -594,7 +594,7 @@ xrealloc_c(void *ptr, size_t nmemb, size_t size,
 #ifdef XALLOC_DEBUG
 	if (debug_level > 0) {
 		struct memory	*mem_exists;
-		ssize_t 	 diff = nsiz - mem->size;
+		ssize_t 	 diff = (ssize_t)(nsiz - mem->size);
 
 		XALLOC_LOCK(xalloc_mutex);
 		xalloc_allocated += diff;
